+++ lib/isolate_contactor.dart
@@ -1,11 +1,5 @@
 library isolate_contactor;
 
-import 'dart:async';
-import 'dart:isolate';
-import 'package:flutter/foundation.dart';
-import 'package:stream_channel/isolate_channel.dart';
-
-part 'src/isolate_contactor.dart';
-part 'src/enum.dart';
-part 'src/internal.dart';
-part 'src/extension.dart';
+export 'src/utils.dart' show ComputeState;
+export 'src/isolate_contactor.dart';
+export 'src/isolate_contactor_controller.dart';