+++ lib/src/enum.dart
@@ -1,4 +1,0 @@
-part of isolate_contactor;
-
-enum _IsolatePort { main, child, debug }
-enum ComputeState { computing, computed }