+++ test/isolate_contactor_test.dart
@@ -1,9 +1,26 @@
-import 'package:flutter/services.dart';
+import 'dart:async';
+
 import 'package:flutter_test/flutter_test.dart';
 import 'package:isolate_contactor/isolate_contactor.dart';
-import 'package:stream_channel/isolate_channel.dart';
 
 void main() {
+  test('Test', () async {
+    StreamController streamController = StreamController.broadcast();
+
+    final stream1 = streamController.stream.listen((event) {
+      print('1: $event');
+    });
+
+    final stream2 = streamController.stream.listen((event) {
+      print('2: $event');
+    });
+
+    streamController.add('adb');
+
+    await Future.delayed(const Duration(seconds: 3));
+    stream1.cancel();
+    stream2.cancel();
+  });
   test('Basic use', () async {
     // Just for waiting till the result has come
     bool valueExit = false;
@@ -23,7 +40,7 @@ void main() {
     // Send 10 to fibonacci isolate function
     isolateContactor.sendMessage(10);
 
-    // Only for waiting the result. Don't need to use in your real app
+    // Only for waiting the result in Console app. Don't need to use in your real app
     while (!valueExit) {
       await Future.delayed(const Duration(milliseconds: 10));
     }
@@ -64,6 +81,7 @@ void main() {
     // Send 20 and 10 to [subtract]
     isolateContactor2.sendMessage([10, 20]);
 
+    // Only for waiting the result in Console app. Don't need to use in your real app
     while (!value1Exit && !value2Exit) {
       await Future.delayed(const Duration(milliseconds: 10));
     }
@@ -88,6 +106,7 @@ void main() {
     // Send 10 and 20 to [isolateFunction]
     isolateContactor.sendMessage([10, 20]);
 
+    // Only for waiting the result in Console app. Don't need to use in your real app
     while (!valueExit) {
       await Future.delayed(const Duration(milliseconds: 10));
     }
@@ -97,10 +116,17 @@ void main() {
 }
 
 dynamic fibonacci(dynamic n) {
-  if (n == 0) return 0;
-  if (n == 1 || n == 2) return 1;
+  if (n <= 2) return 1;
 
-  return fibonacci(n - 2) + fibonacci(n - 1);
+  int n1 = 0, n2 = 1, n3 = 1;
+
+  for (int i = 2; i <= n; i++) {
+    n3 = n1 + n2;
+    n1 = n2;
+    n2 = n3;
+  }
+
+  return n3.round();
 }
 
 // single parameter
@@ -110,14 +136,9 @@ dynamic subtract(dynamic n) => n[1] - n[0];
 dynamic add(dynamic a, dynamic b) => a + b;
 
 // Create your own function here
-void isolateFunction(List<dynamic> params) {
-  final channel = IsolateChannel.connectSend(params.last);
-  channel.stream.listen((rawMessage) {
-    final message = IsolateContactor.getMessage(rawMessage);
-    if (message != null) {
-      // Do more stuff here
-
-      channel.sendResult(add(message[0], message[1]));
-    }
+void isolateFunction(dynamic params) {
+  final channel = IsolateContactorController(params);
+  channel.onIsolateMessage.listen((message) {
+    channel.sendResult(add(message[0], message[1]));
   });
 }